benefici licitacio

+++ database/migrations/2018_03_15_153119_create_licitacions_table.php
@@ -17,7 +17,7 @@ class CreateLicitacionsTable extends Migration
             $table->increments('id')->unsigned();
             $table->integer('id_usuari')->unsigned();
             $table->integer('id_subhasta')->unsigned();
-            $table->integer('preu')->unsigned();
+            $table->decimal('preu',28,10)->unsigned();
             $table->date('temps');
             $table->timestamps();
             $table->foreign('id_usuari')->references('id')->on('users');
